Coding Weather Challenge
This app parses the openweathermap API located at
(http://openweathermap.org/api) to show the current temperature for the
users selected zip codes. It shows the location on a map view, with a
pin having the label of the temperature.

This was my first app that parsed data from an api correctly, so I am
pretty proud of it. It was a coding challenge for a company.

Hope you guys enjoy the work that I put into it. The background image
on the table view is not mine. I found it on google, and I am using it
under the assumption of fair use.